fix: wrap all links in articles in next/link and add target _blank for external links
